% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbConnect,jdbcHiveDriver-method}
\alias{dbConnect,jdbcHiveDriver-method}
\title{method "dbConnect" copied from RJDBC with JDBCConnection replaced with jdbcHiveDriver}
\usage{
\S4method{dbConnect}{jdbcHiveDriver}(drv, url, schema = NULL, user = "", password = "", ...)
}
\description{
method "dbConnect" copied from RJDBC with JDBCConnection replaced with jdbcHiveDriver
}
