% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbGetFields,HiveS2Connection-method}
\alias{dbGetFields,HiveS2Connection-method}
\title{TODO: dbGetFields not working}
\usage{
\S4method{dbGetFields}{HiveS2Connection}(conn, table = "\%", schema = conn@schema_name, ...)
}
\description{
TODO: dbGetFields not working
}
