% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbGetTables,HiveS2Connection-method}
\alias{dbGetTables,HiveS2Connection-method}
\title{TODO: dbGetTables not working}
\usage{
\S4method{dbGetTables}{HiveS2Connection}(conn, table = "\%", schema = conn@schema_name, ...)
}
\description{
TODO: dbGetTables not working
}
