% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbListTables,HS2Connection-method}
\alias{dbListTables,HS2Connection-method}
\title{TODO: why dbListTables shows character(0)}
\usage{
\S4method{dbListTables}{HS2Connection}(conn, pattern = "\%", schema = NULL, ...)
}
\description{
TODO: why dbListTables shows character(0)
}
