% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbQuoteIdentifier,HiveS2Connection,character-method}
\alias{dbQuoteIdentifier,HiveS2Connection,character-method}
\title{method "dbQuoteIdentifier" based on RSQLite}
\usage{
\S4method{dbQuoteIdentifier}{HiveS2Connection,character}(conn, x, ...)
}
\description{
method "dbQuoteIdentifier" based on RSQLite
}
