% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbQuoteIdentifier,HiveS2Connection,SQL-method}
\alias{dbQuoteIdentifier,HiveS2Connection,SQL-method}
\title{method "dbQuoteIdentifier" based on RSQLite
#' @export
setMethod("dbQuoteIdentifier", c("HiveS2Connection", "character"), function(conn, x, ...) {
  if (any(is.na(x))) {
    stop("Cannot pass NA to dbQuoteIdentifier()", call. = FALSE)
  }
  x <- gsub(conn@identifier.quote, paste0(conn@identifier.quote,conn@identifier.quote), enc2utf8(x))
  if (length(x) == 0L) {
    SQL(character(), names = names(x))
  } else {
    SQL(paste(conn@identifier.quote, x, conn@identifier.quote, sep = ""), names = names(x))
  }
})}
\usage{
\S4method{dbQuoteIdentifier}{HiveS2Connection,SQL}(conn, x, ...)
}
\description{
method "dbQuoteIdentifier" based on RSQLite
#' @export
setMethod("dbQuoteIdentifier", c("HiveS2Connection", "character"), function(conn, x, ...) {
  if (any(is.na(x))) {
    stop("Cannot pass NA to dbQuoteIdentifier()", call. = FALSE)
  }
  x <- gsub(conn@identifier.quote, paste0(conn@identifier.quote,conn@identifier.quote), enc2utf8(x))
  if (length(x) == 0L) {
    SQL(character(), names = names(x))
  } else {
    SQL(paste(conn@identifier.quote, x, conn@identifier.quote, sep = ""), names = names(x))
  }
})
}
