% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{HiveS2}
\alias{HiveS2}
\title{function HiveS2 copied from RJDBC
create jdbcHiveDriver dbObj}
\usage{
HiveS2(driverClass = "", classPath = "", identifier.quote = NA)
}
\description{
function HiveS2 copied from RJDBC
create jdbcHiveDriver dbObj
}
