% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Driver.R
\docType{class}
\name{HiveS2Driver-class}
\alias{HiveS2Driver-class}
\title{Driver for HiveServer2}
\description{
Driver for HiveServer2
}
\keyword{HiveS2Driver}
\keyword{class.}
\keyword{driver}
\keyword{internal}
