% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\docType{class}
\name{HS2Connection-class}
\alias{HS2Connection-class}
\title{HS2Connection class inherits from JDBCConnection (RJDBC)}
\description{
HS2Connection class inherits from JDBCConnection (RJDBC)
}
