% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbcHiveDriver.R
\docType{class}
\name{jdbcHiveDriver-class}
\alias{jdbcHiveDriver-class}
\alias{show,jdbcHiveDriver-method}
\title{Driver for HiveServer2}
\usage{
\S4method{show}{jdbcHiveDriver}(object)
}
\description{
Driver for HiveServer2
}
\keyword{class.}
\keyword{driver}
\keyword{internal}
\keyword{jdbcHiveDriver}
