% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{show,HiveS2Connection-method}
\alias{show,HiveS2Connection-method}
\title{#' @rdname HiveS2Connection-class
#' @export}
\usage{
\S4method{show}{HiveS2Connection}(object)
}
\description{
#' @rdname HiveS2Connection-class
#' @export
}
