% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\name{dbGetInfo,HiveS2Connection-method}
\alias{dbGetInfo,HiveS2Connection-method}
\title{jdbcHiveS2Connection info}
\usage{
\S4method{dbGetInfo}{HiveS2Connection}(dbObj, ...)
}
\description{
jdbcHiveS2Connection info
}
