% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\docType{class}
\name{HiveS2Connection-class}
\alias{HiveS2Connection-class}
\title{HiveS2Connection class connection class.
inherits from JDBCConnection (RJDBC)}
\description{
HiveS2Connection class connection class.
inherits from JDBCConnection (RJDBC)
}
\keyword{internal}
