% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R
\docType{class}
\name{HiveS2Connection-class}
\alias{HiveS2Connection-class}
\alias{HiveS2}
\alias{dbQuoteIdentifier,HiveS2Connection,character-method}
\alias{dbQuoteIdentifier,HiveS2Connection,SQL-method}
\alias{dbGetInfo,HiveS2Connection-method}
\alias{show,HiveS2Connection-method}
\alias{dbDisconnect,HiveS2Connection-method}
\alias{dbWriteTable,HiveS2Connection,ANY-method}
\title{HiveS2Connection class connection class.
inherits from JDBCConnection (RJDBC)}
\usage{
HiveS2(driverClass = "", classPath = "", identifier.quote = NA)

\S4method{dbQuoteIdentifier}{HiveS2Connection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{HiveS2Connection,SQL}(conn, x, ...)

\S4method{dbGetInfo}{HiveS2Connection}(dbObj, ...)

\S4method{show}{HiveS2Connection}(object)

\S4method{dbDisconnect}{HiveS2Connection}(conn, ...)

\S4method{dbWriteTable}{HiveS2Connection,ANY}(conn, name, value, ...)
}
\description{
HiveS2Connection class connection class.
inherits from JDBCConnection (RJDBC)

function HiveS2 copied from RJDBC
create jdbcHiveDriver dbObj

jdbcHiveS2Connection info
}
\keyword{internal}
