% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiveS2Connection.R, R/HiveS2Connector.R,
%   R/dbListTables.R
\name{HiveS2}
\alias{HiveS2}
\alias{dbQuoteIdentifier,HiveS2Connection,character-method}
\alias{dbQuoteIdentifier,HiveS2Connection,SQL-method}
\alias{dbGetInfo,HiveS2Connection-method}
\alias{show,HiveS2Connection-method}
\alias{dbDisconnect,HiveS2Connection-method}
\alias{dbWriteTable,HiveS2Connection,ANY-method}
\alias{dbConnect,HiveS2Driver-method}
\alias{dbListTables,HiveS2Connection-method}
\title{function HiveS2 copied from RJDBC
creates HiveS2Driver dbObj}
\usage{
HiveS2(driverClass = "", classPath = "", identifier.quote = "`")

\S4method{dbQuoteIdentifier}{HiveS2Connection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{HiveS2Connection,SQL}(conn, x, ...)

\S4method{dbGetInfo}{HiveS2Connection}(dbObj, ...)

\S4method{show}{HiveS2Connection}(object)

\S4method{dbDisconnect}{HiveS2Connection}(conn, ...)

\S4method{dbWriteTable}{HiveS2Connection,ANY}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  force = FALSE,
  ...
)

\S4method{dbConnect}{HiveS2Driver}(drv, host, port, schema = NULL, user = "", password = "", ...)

\S4method{dbListTables}{HiveS2Connection}(conn, pattern = NULL, schema = NULL, ...)
}
\arguments{
\item{overwrite}{allows to specify weather existing table with the same name should be removed}

\item{drv}{HiveS2Driver object}

\item{schema}{- database which tables will be listed from}

\item{pattern}{- specify pattern for SQL 'SHOW TABLES LIKE' statment}
}
\description{
function HiveS2 copied from RJDBC
creates HiveS2Driver dbObj

dbWriteTable writtes table into local temporary dile and than it is laded into Hive with 'LOAD DATA' statment
Sceleton of the fuction adapted from RJDBC packade

when specifying schema all queries that follow will be executed on that schema (warning message)
}
